rag-engine/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
requests
